+++ grid2op/tests/test_PandaPowerBackend.py
@@ -16,10 +16,18 @@ from grid2op.tests.helper_path_test import PATH_DATA_TEST_PP, PATH_DATA_TEST
 from grid2op.Backend import PandaPowerBackend
 
 from grid2op.tests.helper_path_test import HelperTests
-from grid2op.tests.BaseBackendTest import BaseTestNames, BaseTestLoadingCase, BaseTestLoadingBackendFunc
-from grid2op.tests.BaseBackendTest import BaseTestTopoAction, BaseTestEnvPerformsCorrectCascadingFailures
-from grid2op.tests.BaseBackendTest import BaseTestChangeBusAffectRightBus, BaseTestShuntAction
-from grid2op.tests.BaseBackendTest import BaseTestResetEqualsLoadGrid, BaseTestVoltageOWhenDisco, BaseTestChangeBusSlack
+from grid2op.tests.BaseBackendTest import BaseTestNames
+from grid2op.tests.BaseBackendTest import BaseTestLoadingCase
+from grid2op.tests.BaseBackendTest import BaseTestLoadingBackendFunc
+from grid2op.tests.BaseBackendTest import BaseTestTopoAction
+from grid2op.tests.BaseBackendTest import BaseTestEnvPerformsCorrectCascadingFailures
+from grid2op.tests.BaseBackendTest import BaseTestChangeBusAffectRightBus
+from grid2op.tests.BaseBackendTest import BaseTestShuntAction
+from grid2op.tests.BaseBackendTest import BaseTestResetEqualsLoadGrid
+from grid2op.tests.BaseBackendTest import BaseTestVoltageOWhenDisco
+from grid2op.tests.BaseBackendTest import BaseTestChangeBusSlack
+from grid2op.tests.BaseBackendTest import BaseIssuesTest
+from grid2op.tests.BaseBackendTest import BaseStatusActions
 PATH_DATA_TEST_INIT = PATH_DATA_TEST
 PATH_DATA_TEST = PATH_DATA_TEST_PP
 
@@ -126,6 +134,16 @@ class TestChangeBusSlack(HelperTests, BaseTestChangeBusSlack):
         return PandaPowerBackend(detailed_infos_for_cascading_failures=detailed_infos_for_cascading_failures)
 
 
+class TestIssuesTest(HelperTests, BaseIssuesTest):
+    def make_backend(self, detailed_infos_for_cascading_failures=False):
+        return PandaPowerBackend(detailed_infos_for_cascading_failures=detailed_infos_for_cascading_failures)
+
+
+class TestStatusAction(HelperTests, BaseStatusActions):
+    def make_backend(self, detailed_infos_for_cascading_failures=False):
+        return PandaPowerBackend(detailed_infos_for_cascading_failures=detailed_infos_for_cascading_failures)
+
+
 # Specific to pandapower power
 class TestChangeBusAffectRightBus2(unittest.TestCase):
     def skip_if_needed(self):
@@ -179,6 +197,7 @@ class TestChangeBusAffectRightBus2(unittest.TestCase):
         assert np.all(np.isfinite(obs.v_or))
         assert np.sum(env.backend._grid["bus"]["in_service"]) == 14
         assert env.backend._grid["trafo"]["hv_bus"][2] == 4
+
 
 if __name__ == "__main__":
     unittest.main()